aed_iot_security/requirements.txt
streamlit
mysql-connector-python
pandas
numpy
plotly
streamlit-geolocation
